hw5pr3: complete

+++ hw/hw5pr3/mtest.c
@@ -10,7 +10,9 @@ const char *HELP_MESSAGE =
     "%1$s - test mmap features\n"
     "\n"
     "USAGE:\n"
-    "\t%1$s test_number\n";
+    "\t%1$s test_number\n"
+    "AVAILABLE TESTS:\n"
+    "\t{1,2,3,4}\n";
 
 void handler(int sig) {
   printf("Signal [%i: %s] received\n", sig, strsignal(sig));
@@ -22,7 +24,8 @@ int write_helper(int fd, const void* buf, size_t count) {
   for (size_t written = 0; written < count;) {
     ssize_t r = write(fd, buf + written, count - written);
     if (r == -1) {
-      return errno;
+      fprintf(stderr, "ERROR: Failed writing to file: %s\n", strerror(errno));
+      exit(255);
     }
     written += r;
   }
@@ -31,7 +34,12 @@ int write_helper(int fd, const void* buf, size_t count) {
 
 int test1() {
   printf("Executing Test #1 (write to r/o mmap):\n");
-  char *map = mmap(NULL, sysconf(_SC_PAGE_SIZE), PROT_READ, MAP_ANONYMOUS|MAP_PRIVATE, 0, 0);
+  const int PAGE_SIZE = sysconf(_SC_PAGE_SIZE);
+  char *map = mmap(NULL, PAGE_SIZE, PROT_READ, MAP_ANONYMOUS|MAP_PRIVATE, 0, 0);
+  if (map == MAP_FAILED) {
+    fprintf(stderr, "ERROR: Failed to call mmap for anonymous region: %s\n", strerror(errno));
+    exit(255);
+  }
   // TODO: should we set this to something before?
   printf("map[3] == '%c'\n", map[3]);
   char c = 'B';
@@ -46,6 +54,10 @@ int test1() {
 int mktempfile() {
   char template[] = "/tmp/gjmtest-XXXXXX";
   int fd = mkstemp(template);
+  if (fd == -1) {
+    fprintf(stderr, "ERROR: Failed to create temporary file: %s\n", strerror(errno));
+    exit(255);
+  }
   unlink(template);
   return fd;
 }
@@ -54,7 +66,12 @@ int test2() {
   printf("Executing Test #2 (write to MAP_SHARED):\n");
   int fd = mktempfile();
   write_helper(fd, "1234567890", 10);
-  char *map = mmap(NULL, sysconf(_SC_PAGE_SIZE), PROT_READ|PROT_WRITE, MAP_SHARED, fd, 0);
+  const int PAGE_SIZE = sysconf(_SC_PAGE_SIZE);
+  char *map = mmap(NULL, PAGE_SIZE, PROT_READ|PROT_WRITE, MAP_SHARED, fd, 0);
+  if (map == MAP_FAILED) {
+    fprintf(stderr, "ERROR: Failed to call mmap for temporary file: %s\n", strerror(errno));
+    exit(255);
+  }
   printf("map[3] == '%c'\n", map[3]);
   char c = 'B';
   printf("writing a '%c'\n", c);
@@ -76,7 +93,12 @@ int test3() {
   printf("Executing Test #3 (write to MAP_PRIVATE):\n");
   int fd = mktempfile();
   write_helper(fd, "1234567890", 10);
-  char *map = mmap(NULL, sysconf(_SC_PAGE_SIZE), PROT_READ|PROT_WRITE, MAP_PRIVATE, fd, 0);
+  const int PAGE_SIZE = sysconf(_SC_PAGE_SIZE);
+  char *map = mmap(NULL, PAGE_SIZE, PROT_READ|PROT_WRITE, MAP_PRIVATE, fd, 0);
+  if (map == MAP_FAILED) {
+    fprintf(stderr, "ERROR: Failed to call mmap for temporary file: %s\n", strerror(errno));
+    exit(255);
+  }
   printf("map[3] == '%c'\n", map[3]);
   char c = 'B';
   printf("writing a '%c'\n", c);
@@ -92,6 +114,53 @@ int test3() {
     printf("written value is not visible\n");
     return 1;
   }
+
+  return 0;
+}
+
+int test4() {
+  printf("Executing Test #4 (write into not yet created part of file):\n");
+  int fd = mktempfile();
+  const int PAGE_SIZE = sysconf(_SC_PAGE_SIZE);
+
+  const int FILE_LENGTH = PAGE_SIZE + 5;
+
+  printf("Writing %i bytes of data into temporary file\n", FILE_LENGTH);
+  char *buf = malloc(FILE_LENGTH);
+  memset(buf, 'A', FILE_LENGTH);
+  write_helper(fd, buf, FILE_LENGTH);
+  
+  char *map = mmap(NULL, PAGE_SIZE * 2, PROT_READ|PROT_WRITE, MAP_SHARED, fd, 0);
+  if (map == MAP_FAILED) {
+    fprintf(stderr, "ERROR: Failed to call mmap for temporary file: %s\n", strerror(errno));
+    exit(255);
+  }
+  // Verify memory is all zero beyond defined area.
+  for (int i = FILE_LENGTH; i < PAGE_SIZE * 2; i++) {
+    if (map[i] != '\0') {
+      printf("Memory beyond written file is not all zero\n");
+      return 2;
+    }
+  }
+  printf("Verified that memory beyond written file is all zero\n");
+
+
+  map[FILE_LENGTH] = 'X';
+  lseek(fd, FILE_LENGTH + 16, SEEK_SET);
+  write(fd, "B", 1);
+
+  lseek(fd, FILE_LENGTH, SEEK_SET);
+  read(fd, buf, 1);
+
+  if (buf[0] == 'X') {
+    printf("The byte can be seen\n");
+    return 0;
+  } else {
+    printf("The byte cannot be seen\n");
+    return 1;
+  }
+
+  return 0;
 }
 
 int main(int argc, char **argv) {
@@ -127,6 +196,11 @@ int main(int argc, char **argv) {
   // Test 3
   if (argv[1][0] == '3') {
     return test3();
+  }
+
+  // Test 4
+  if (argv[1][0] == '4') {
+    return test4();
   }
 
   printf("Did not recognize given test number. Run \"%s help\" for usage information\n", argv[0]);
